Reworks grid generation so every puzzle ID yields a valid puzzle.
We used to count the number of words in the grid separately from the generation process, and just move on to the next ID if the grid had too few words.  Now we just reuse the RNG and make a new grid, as many times as we need to get at least 50 words.